fix(decompose): move 3 test-helper modules into __tests__/
Previous commit (60577c28) renamed three helpers from <name>.test.ts to
<name>.ts to stop vitest globbing them as tests. That fixed the vitest
exit-1 on "No test suite found", but reclassified them as PRODUCTION
SOURCE (no .test.ts suffix, not under /__tests__/), which:

  - Pushed aggregate BCI from 50.60 to 50.65 (over budget 50.61) because
    each file imports @vt/graph-model / @vt/graph-db-server / @vt/graph-state
    types, widening cross-package boundaries.
  - Triggered package-boundaries.test forbidden-import on
    folder-loading.test/test-support.ts:11 (imports
    @vt/graph-db-server/state/graph-store outside allowlisted paths).

The fix: move them under sibling __tests__/ directories, which
import-graph.ts:90 and package-boundaries.test.ts both exclude. Parent
.test.ts files updated to import via __tests__/<name>.

Local verification:
  - hypergraph-complexity.test.ts: BCI 50.60 (budget 50.61) ✓
  - package-boundaries.test.ts: 4/4 pass ✓
  - CodeMirrorEditorView.test.ts: 44 pass / 4 skip ✓
