+++ webapp/src/shell/UI/floating-windows/editors/CodeMirrorEditorView.test/__tests__/imagePasteHandler.ts
